consumer/requirements.txt
nats-py==2.4.0
clickhouse-driver==0.2.6